requirements.txt
absl-py==0.3.0
astor==0.7.1
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==2.0.7
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
idna==2.7
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.15.0
oauthlib==2.1.0
Pillow==5.2.0
protobuf==3.6.0
psycopg2==2.7.5
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
scipy==1.1.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tensorboard==1.9.0
tensorflow==1.9.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
absl-py==0.3.0
astor==0.7.1
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==2.0.7
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
idna==2.7
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.15.0
oauthlib==2.1.0
Pillow==5.2.0
protobuf==3.6.0
psycopg2==2.7.5
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
scipy==1.1.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tensorboard==1.9.0
tensorflow==1.9.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
